fix boolean issue

+++ environments.py
@@ -220,7 +220,7 @@ class RodentSingleClipTrack(PipelineEnv):
     info['cur_frame'] += 1
 
     done = termination_error > self._termination_threshold
-    done = jp.array(done, int)
+    done = jp.array(done, float)
 
     state.metrics.update(
         rcom=rcom,
